Partial rebuild of KarKam app in branch 33-implement-base-settings-page-functionality-using-get_it.
Button functionality works correctly.

Settings page functionality doesn't NOT work due to _KarKam being a StatelessWidget by deliberate choice.
